# api/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private REST API to for user interface application",
        "title": "storePredictor private REST API",
        "contact": {
            "name": "Ales Jandera",
            "url": "http://storepredictor.com",
            "email": "ales@storepredictor.com"
        },
        "license": {
            "name": "Commercial"
        },
        "version": "1.0"
    },
    "host": "localhost:9996",
    "basePath": "/",
    "paths": {
...
        "/accounts/{lang}": {
            "post": {
                "description": "Endpoint to StoreWeights data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StoreWeights"
                ],
                "summary": "Endpoint to store StoreWeights data",
                "parameters": [
                    {
                        "description": "StoreWeights object to store in database",
                        "name": "sw.StoreWeights",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.StoreWeights"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/store-weights": {
            "put": {
                "description": "Endpoint to update StoreWeights data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StoreWeights"
                ],
                "summary": "Endpoint to update StoreWeights data",
                "parameters": [
                    {
                        "description": "StoreWeights object to update in database",
                        "name": "sw.StoreWeights",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.StoreWeights"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
        "/stores": {
            "get": {
                "description": "Endpoint to get stores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stores"
                ],
                "summary": "Endpoint to get stores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/structs.Stores"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Endpoint to update Stores data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stores"
                ],
                "summary": "Endpoint to update Stores data",
                "parameters": [
                    {
                        "description": "Stores object to update in database",
                        "name": "s.Stores",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.Stores"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint to store Stores data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stores"
                ],
                "summary": "Endpoint to store Stores data",
                "parameters": [
                    {
                        "description": "Stores object to store in database",
                        "name": "s.Stores",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.Stores"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "structs.StoreWeights": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                },
                "beta": {
                    "type": "number"
                },
                "c": {
                    "type": "number"
                },
                "d": {
                    "type": "number"
                },
                "delta": {
                    "type": "number"
                },
                "e": {
                    "type": "number"
                },
                "gama": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "probabilityWeights": {
                    "type": "string"
                },
                "storeRefer": {
                    "type": "string"
                }
            }
        },
        "structs.Stores": {
            "type": "object",
            "properties": {
                "accountRefer": {
                    "type": "string"
                },
                "actualCustomerSatisfaction": {
                    "type": "number"
                },
                "actualStorePower": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "countryCode": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastPrediction": {
                    "type": "string"
                },
                "maximalProductPrice": {
                    "type": "number"
                },
                "minimalProductPrice": {
                    "type": "number"
                },
                "perceivedValue": {
                    "type": "number"
                },
                "productSell": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "utils.Res": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}